backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.2
motor==3.1.1
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.2
cloudinary==1.33.0
reportlab==4.0.1
tensorflow==2.15.0
numpy==1.26.0
scikit-learn==1.3.3
roboflow==1.1.7
requests==2.32.0
